refactor: if to install

+++ install.sh
@@ -39,28 +39,29 @@ update_mondoc() {
     sudo mv mondoc /usr/local/bin/mondoc
     check_command_success
     echo "Mondoc.ai update and configuration completed successfully! Use 'mondoc' to start the Mondoc.ai CLI."
-    return
 }
 
 if [ -f "/usr/local/bin/mondoc" ]; then
     echo "Mondoc is already installed. Updating..."
     update_mondoc
 fi
-echo "Installing Mondoc for the first time..."
-# Passo 1: Download the binary
-echo "Downloading Mondoc binary..."
-curl -O https://raw.githubusercontent.com/WalysonGO/mondoc.ai/master/mondoc > /dev/null 2>&1
-check_command_success
 
-# Passo 2: Conceder permissão de execução ao binário
-echo "Granting execution permission..."
-chmod +x mondoc
-check_command_success
+if [ ! -f "/usr/local/bin/mondoc" ]; then
+    echo "Installing Mondoc for the first time..."
+    # Passo 1: Download the binary
+    echo "Downloading Mondoc binary..."
+    curl -O https://raw.githubusercontent.com/WalysonGO/mondoc.ai/master/mondoc > /dev/null 2>&1
+    check_command_success
 
-# Passo 3: Mover o binário para o diretório de execução global
-echo "Moving binary to /usr/local/bin..."
-sudo mv mondoc /usr/local/bin/mondoc
-check_command_success
+    # Passo 2: Conceder permissão de execução ao binário
+    echo "Granting execution permission..."
+    chmod +x mondoc
+    check_command_success
 
-echo "Mondoc.ai installation and configuration completed successfully! Use 'mondoc' to start the Mondoc.ai CLI."
-return
+    # Passo 3: Mover o binário para o diretório de execução global
+    echo "Moving binary to /usr/local/bin..."
+    sudo mv mondoc /usr/local/bin/mondoc
+    check_command_success
+
+    echo "Mondoc.ai installation and configuration completed successfully! Use 'mondoc' to start the Mondoc.ai CLI."
+fi
